apps/backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
razorpay==1.4.1
boto3==1.34.0
python-dotenv==1.0.0
apscheduler==3.10.4
python-socketio==5.10.0
aiosmtplib==3.0.1
jinja2==3.1.2
httpx==0.25.2
